separate annotation from processor
