+++ annotation/build.gradle
@@ -2,11 +2,6 @@ apply plugin: 'java-library'
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-
-    //add interface reference to be processed
-    compile project(':annotation')
-
-    compile 'com.google.auto.service:auto-service:1.0-rc3'
 }
 
 sourceCompatibility = "1.7"
