+++ annotation/src/main/java/com/ferbajoo/annotation/Foo.java
@@ -1,6 +1,7 @@
-package com.mavi.processor;
+package com.ferbajoo.annotation;
 
 import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
@@ -12,6 +13,7 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
+@Inherited()
 public @interface Foo {
     String name() default "";
     String value() default  "";
