+++ app/build.gradle
@@ -32,6 +32,9 @@ dependencies {
     compile 'com.android.support:recyclerview-v7:26.1.0'
     implementation 'com.android.support:appcompat-v7:26.1.0'
     implementation 'com.android.support.constraint:constraint-layout:1.0.2'
+    //interface references, this must be included in the project
+    compile project(path: ':annotation')
+    //annotation processor, this is not included in the project
     provided project(':processor')
     annotationProcessor project(':processor')
 }
