+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':processor'
+include ':app', ':processor', ':annotation'
